# app/Http/Controllers/CartController.php
<?php

namespace App\Http\Controllers;

use App\Models\Cart;
use App\Models\Product;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Auth;

/**
 * @OA\Tag(
 *     name="Cart",
 *     description="API Endpoints for cart management"
 * )
 */
class CartController extends Controller
{
    /**
     * @OA\Post(
     *     path="/api/cart/add",
     *     summary="Add a product to a user's cart",
     *     tags={"Cart"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"user_id", "product_id", "quantity"},
     *             @OA\Property(
     *                 property="user_id",
     *                 type="string",
     *                 format="uuid",
     *                 description="UUID of the user",
     *                 example="905932a9-dbaf-49cd-bca9-451914c71d19"
     *             ),
     *             @OA\Property(
     *                 property="product_id",
     *                 type="string",
     *                description="UUID of the product to add",
     *                 example="905932a9-dbaf-49cd-bca9-451914c71d19"
     *             ),
     *             @OA\Property(
     *                 property="quantity",
     *                 type="integer",
     *                 description="Quantity of the product to add",
     *                 minimum=1,
     *                 example=1
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Product added to cart successfully",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true),
     *             @OA\Property(property="message", type="string", example="Product added to cart successfully"),
     *             @OA\Property(
     *                 property="data",
     *                 type="object",
     *                 @OA\Property(property="user_id", type="string", format="uuid", example="905932a9-dbaf-49cd-bca9-451914c71d19"),
     *                 @OA\Property(property="product_id", type="string", example="string"),
     *                 @OA\Property(property="quantity", type="integer", example=1),
     *                 @OA\Property(
     *                     property="product",
     *                     type="object",
     *                     @OA\Property(property="id", type="string", example="string"),
     *                     @OA\Property(property="name", type="string", example="Product Name"),
     *                     @OA\Property(property="price", type="number", format="float", example=99.99),
     *                     @OA\Property(property="image_url", type="string", example="https://example.com/image.jpg")
     *                 )
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="The user id field is required.")
     *         )
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Server error",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="Failed to add product to cart")
     *         )
     *     )
     * )
     */
    public function addToCart(Request $request)
    {
        $request->validate([
            'user_id' => 'required|string|size:36|exists:users,id',
            'product_id' => 'required|exists:products,id',
            'quantity' => 'required|integer|min:1',
        ]);

        try {
            $product = Product::findOrFail($request->product_id);

            // Check if the product is already in the cart
            $cartItem = Cart::where('user_id', $request->user_id)
                ->where('product_id', $product->id)
                ->first();

            if ($cartItem) {
                // Update quantity if product already exists in cart
                $cartItem->update([
                    'quantity' => $cartItem->quantity + $request->quantity
                ]);
            } else {
                // Create new cart item
                $cartItem = Cart::create([
                    'user_id' => $request->user_id,
                    'product_id' => $product->id,
                    'quantity' => $request->quantity
                ]);
            }

            // Get updated cart item with product details
            $cartItem = Cart::with('product')
                ->where('user_id', $request->user_id)
                ->where('product_id', $product->id)
                ->first();

                $trafficSource = new \App\Models\TrafficSource();
                $trafficSource->source = 'add_to_cart';
                $trafficSource->visits = 1;
                $trafficSource->orders = 0;
                $trafficSource->recorded_at = now();
                $trafficSource->save();

            return response()->json([
                'success' => true,
                'message' => 'Product added to cart successfully',
                'data' => $cartItem
            ]);

        } catch (\Exception $e) {
            return response()->json([
                'success' => false,
                'message' => 'Failed to add product to cart: ' . $e->getMessage()
            ], 500);
        }
    }

    /**
     * @OA\Get(
     *     path="/api/cart",
     *     summary="Get a user's cart items",
     *     tags={"Cart"},
     *     @OA\Parameter(
     *         name="user_id",
     *         in="query",
     *         required=true,
     *         description="ID of the user",
     *         @OA\Schema(type="string")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Cart items retrieved successfully",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true),
     *             @OA\Property(
     *                 property="data",
     *                 type="array",
     *                 @OA\Items(
     *                     type="object",
     *                     @OA\Property(property="user_id", type="string", example="agsjad-dfhbd-saadsd"),
     *                     @OA\Property(property="product_id", type="string", example="hjas-jkdsf"),
     *                     @OA\Property(property="quantity", type="integer", example=1),
     *                     @OA\Property(
     *                         property="product",
     *                         type="object",
     *                         @OA\Property(property="id", type="string", example="ahjdsdkj"),
     *                         @OA\Property(property="name", type="string", example="Product Name"),
     *                         @OA\Property(property="price", type="number", format="float", example=99.99),
     *                         @OA\Property(property="image_url", type="string", example="https://example.com/image.jpg")
     *                     )
     *                 )
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="The user id field is required.")
     *         )
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Server error",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="Failed to get cart items")
     *         )
     *     )
     * )
     */
    public function getCart(Request $request)
    {
        $request->validate([
            'user_id' => 'required|exists:users,id'
        ]);

        try {
            $cartItems = Cart::with('product')
                ->where('user_id', $request->user_id)
                ->get();

                $trafficSource = new \App\Models\TrafficSource();
                $trafficSource->source = 'get_cart';
                $trafficSource->visits = 1;
                $trafficSource->orders = 0;
                $trafficSource->recorded_at = now();
                $trafficSource->save();

            return response()->json([
                'success' => true,
                'data' => $cartItems
            ]);

        } catch (\Exception $e) {
            return response()->json([
                'success' => false,
                'message' => 'Failed to get cart items: ' . $e->getMessage()
            ], 500);
        }
    }

    /**
     * @OA\Patch(
     *     path="/api/cart/update",
     *     summary="Update the quantity of a product in a user's cart",
     *     tags={"Cart"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"user_id", "product_id", "quantity"},
     *             @OA\Property(
     *                 property="user_id",
     *                 type="string",
     *                 description="ID of the user",
     *                 example=1
     *             ),
     *             @OA\Property(
     *                 property="product_id",
     *                 type="string",
     *                 description="ID of the product in the cart",
     *                 example=1
     *             ),
     *             @OA\Property(
     *                 property="quantity",
     *                 type="integer",
     *                 description="New quantity of the product",
     *                 minimum=1,
     *                 example=2
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Cart item updated successfully",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true),
     *             @OA\Property(property="message", type="string", example="Cart item updated successfully"),
     *             @OA\Property(
     *                 property="data",
     *                 type="object",
     *                 @OA\Property(property="user_id", type="string", example="string"),
     *                 @OA\Property(property="product_id", type="string", example="string"),
     *                 @OA\Property(property="quantity", type="integer", example=2),
     *                 @OA\Property(
     *                     property="product",
     *                     type="object",
     *                     @OA\Property(property="id", type="string", example="string"),
     *                     @OA\Property(property="name", type="string", example="Product Name"),
     *                     @OA\Property(property="price", type="number", format="float", example=99.99),
     *                     @OA\Property(property="image_url", type="string", example="https://example.com/image.jpg")
     *                 )
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Cart item not found",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="Cart item not found")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="The user id field is required.")
     *         )
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Server error",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="Failed to update cart item")
     *         )
     *     )
     * )
     */
    public function updateCartItem(Request $request)
    {
        $request->validate([
            'user_id' => 'required|exists:users,id',
            'product_id' => 'required|exists:products,id',
            'quantity' => 'required|integer|min:1',
        ]);

        try {
            $cartItem = Cart::where('user_id', $request->user_id)
                ->where('product_id', $request->product_id)
                ->first();

            if (!$cartItem) {
                return response()->json([
                    'success' => false,
                    'message' => 'Cart item not found'
                ], 404);
            }

            $cartItem->update([
                'quantity' => $request->quantity
            ]);

            // Get updated cart item with product details
            $cartItem = Cart::with('product')
                ->where('user_id', $request->user_id)
                ->where('product_id', $request->product_id)
                ->first();

                $trafficSource = new \App\Models\TrafficSource();
        $trafficSource->source = 'update_cart_item';
        $trafficSource->visits = 1;
        $trafficSource->orders = 0;
        $trafficSource->recorded_at = now();
        $trafficSource->save();

            return response()->json([
                'success' => true,
                'message' => 'Cart item updated successfully',
                'data' => $cartItem
            ]);

        } catch (\Exception $e) {
            return response()->json([
                'success' => false,
                'message' => 'Failed to update cart item: ' . $e->getMessage()
            ], 500);
        }
    }

    /**
     * @OA\Delete(
     *     path="/api/cart/remove",
     *     summary="Remove a product from a user's cart",
     *     tags={"Cart"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"user_id", "product_id"},
     *             @OA\Property(
     *                 property="user_id",
     *                 type="string",
     *                 description="ID of the user",
     *                 example=1
     *             ),
     *             @OA\Property(
     *                 property="product_id",
     *                 type="string",
     *                 description="ID of the product to remove from cart",
     *                 example="string"
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Cart item removed successfully",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true),
     *             @OA\Property(property="message", type="string", example="Cart item removed successfully")
     *         )
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Cart item not found",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="Cart item not found")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="The user id field is required.")
     *         )
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Server error",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="Failed to remove cart item")
     *         )
     *     )
     * )
     */
    public function removeCartItem(Request $request)
    {
        $request->validate([
            'user_id' => 'required|exists:users,id',
            'product_id' => 'required|exists:products,id',
        ]);

        try {
            $cartItem = Cart::where('user_id', $request->user_id)
                ->where('product_id', $request->product_id)
                ->first();

            if (!$cartItem) {
                return response()->json([
                    'success' => false,
                    'message' => 'Cart item not found'
                ], 404);
            }

            $cartItem->delete();

            $trafficSource = new \App\Models\TrafficSource();
        $trafficSource->source = 'delete_cart_item';
        $trafficSource->visits = 1;
        $trafficSource->orders = 0;
        $trafficSource->recorded_at = now();
        $trafficSource->save();

            return response()->json([
                'success' => true,
                'message' => 'Cart item removed successfully'
            ]);

        } catch (\Exception $e) {
            return response()->json([
                'success' => false,
                'message' => 'Failed to remove cart item: ' . $e->getMessage()
            ], 500);
        }
    }
}