# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cartmate API Documentation",
        "description": "API documentation for Cartmate Backend",
        "contact": {
            "email": "admin@cartmate.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Cartmate API Server"
        }
    ],
    "paths": {
...
        "/api/cart/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add a product to a user's cart",
                "operationId": "af5c3a2c065e9b55aaa49e4cf9b0f9ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "UUID of the user",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "905932a9-dbaf-49cd-bca9-451914c71d19"
                                    },
                                    "product_id": {
                                        "description": "UUID of the product to add",
                                        "type": "string",
                                        "example": "905932a9-dbaf-49cd-bca9-451914c71d19"
                                    },
                                    "quantity": {
                                        "description": "Quantity of the product to add",
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product added to cart successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product added to cart successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "905932a9-dbaf-49cd-bca9-451914c71d19"
                                                },
                                                "product_id": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "product": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "string"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Product Name"
                                                        },
                                                        "price": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 99.99
                                                        },
                                                        "image_url": {
                                                            "type": "string",
                                                            "example": "https://example.com/image.jpg"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The user id field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to add product to cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get a user's cart items",
                "operationId": "399a5f817b36b0db308a434d07715bad",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cart items retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "type": "string",
                                                        "example": "agsjad-dfhbd-saadsd"
                                                    },
                                                    "product_id": {
                                                        "type": "string",
                                                        "example": "hjas-jkdsf"
                                                    },
                                                    "quantity": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "product": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "example": "ahjdsdkj"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Product Name"
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 99.99
                                                            },
                                                            "image_url": {
                                                                "type": "string",
                                                                "example": "https://example.com/image.jpg"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The user id field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to get cart items"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token Based Authentication",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}